feat: Создть новые классы и интерфейсы. Внести соответствующие изменения в другие классы
