+++ src/main/java/ru/yandex/practicum/filmorate/services/FilmService.java
@@ -1,0 +1,17 @@
+package ru.yandex.practicum.filmorate.services;
+
+import org.springframework.stereotype.Service;
+
+import java.util.Set;
+
+@Service
+public class FilmService {
+//    будет отвечать за операции с фильмами,
+//    — добавление лайк,
+//    - удаление лайка,
+//    - вывод 10 наиболее популярных фильмов по количеству лайков.
+
+//    Пусть пока каждый пользователь может поставить лайк фильму только один раз.
+
+
+}
