+++ src/main/java/ru/yandex/practicum/filmorate/storages/UserStorage.java
@@ -1,0 +1,15 @@
+package ru.yandex.practicum.filmorate.storages;
+
+import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.List;
+
+public interface UserStorage {
+
+    User addUser(User user);
+
+    User updateUser(User user);
+
+    List<User> getUsers();
+}
